+++ frontend/catalyst/pennylane_extensions.py
@@ -72,7 +72,11 @@ from catalyst.jax_tracer import (
     trace_quantum_tape,
     unify_result_types,
 )
-from catalyst.utils.contexts import EvaluationContext, EvaluationMode, JaxTracingContext
+from catalyst.tracing.contexts import (
+    EvaluationContext,
+    EvaluationMode,
+    JaxTracingContext,
+)
 from catalyst.utils.exceptions import CompileError, DifferentiableCompileError
 from catalyst.utils.jax_extras import (
     ClosedJaxpr,
@@ -269,7 +273,8 @@ class QJITDevice(qml.QubitDevice):
             QJITDevice.operations += ["Adjoint"]
 
         if QJITDevice._check_quantum_control(config):  # pragma: nocover
-            QJITDevice.operations += ["QCtrl"]
+            # TODO: Once control is added on the frontend.
+            ...
 
     @staticmethod
     def _set_supported_observables(config):
@@ -526,22 +531,18 @@ def _check_grad_params(
 def _unflatten_derivatives(results, out_tree, argnum, num_results):
     """Unflatten the flat list of derivatives results given the out tree."""
     num_trainable_params = len(argnum) if isinstance(argnum, list) else 1
-    results = tuple(
-        tuple(results[i * num_trainable_params : i * num_trainable_params + num_trainable_params])
-        for i in range(0, num_results)
-    )
-    # In jax,  argnums=[int] wraps single derivatives in a tuple
-    if (
-        (isinstance(argnum, int) or argnum is None)
-        and num_trainable_params == 1
-        and num_results != 1
-    ):
-        results = tuple(r[0] if len(r) == 1 else r for r in results)
-    if out_tree.children() != []:
-        results = tree_unflatten(out_tree, results)
-    else:
-        results = results[0]
-    return results
+    results_final = []
+    for i in range(0, num_results):
+        intermediate_results = results[
+            i * num_trainable_params : i * num_trainable_params + num_trainable_params
+        ]
+        if not (isinstance(argnum, int) or argnum is None):
+            intermediate_results = tuple(intermediate_results)
+        else:
+            intermediate_results = intermediate_results[0]
+        results_final.append(intermediate_results)
+    results_final = tree_unflatten(out_tree, results_final)
+    return results_final
 
 
 class Grad:
@@ -598,7 +599,7 @@ def grad(f: DifferentiableLike, *, method=None, h=None, argnum=None):
 
     This function allows the gradient of a hybrid quantum-classical function to be computed within
     the compiled program. Outside of a compiled function, this function will simply dispatch to its
-    JAX counterpart ``jax.grad``.
+    JAX counterpart ``jax.grad``. The function ``f`` can return any pytree-like shape.
 
     .. warning::
 
@@ -728,7 +729,7 @@ def jacobian(f: DifferentiableLike, *, method=None, h=None, argnum=None):
 
     This function allows the Jacobian of a hybrid quantum-classical function to be computed within
     the compiled program. Outside of a compiled function, this function will simply dispatch to its
-    JAX counterpart ``jax.jacobian``.
+    JAX counterpart ``jax.jacobian``. The function ``f`` can return any pytree-like shape.
 
     Args:
         f (Callable): a function or a function object to differentiate
@@ -781,8 +782,8 @@ def jacobian(f: DifferentiableLike, *, method=None, h=None, argnum=None):
             return g(x)
 
     >>> workflow(jnp.array([2.0, 1.0]))
-    array([[-1.32116540e-07,  1.33781874e-07],
-           [-4.20735506e-01,  4.20735506e-01]])
+    array([[ 3.48786850e-16 -4.20735492e-01]
+           [-8.71967125e-17  4.20735492e-01]])
     """
     scalar_out = False
     return Grad(f, GradParams(method, scalar_out, h, argnum))
@@ -794,7 +795,8 @@ def jvp(f: DifferentiableLike, params, tangents, *, method=None, h=None, argnum=
 
     This function allows the Jacobian-vector Product of a hybrid quantum-classical function to be
     computed within the compiled program. Outside of a compiled function, this function will simply
-    dispatch to its JAX counterpart ``jax.jvp``.
+    dispatch to its JAX counterpart ``jax.jvp``. The function ``f`` can return any pytree-like
+    shape.
 
     Args:
         f (Callable): Function-like object to calculate JVP for
@@ -807,8 +809,8 @@ def jvp(f: DifferentiableLike, params, tangents, *, method=None, h=None, argnum=
         h (float): the step-size value for the finite-difference (``"fd"``) method
         argnum (Union[int, List[int]]): the params' indices to differentiate.
 
-    Returns (Tuple[Array]):
-        Return values of ``f`` paired with the JVP values.
+    Returns:
+        Tuple[Any]: Return values of ``f`` paired with the JVP values.
 
     Raises:
         TypeError: invalid parameter types
@@ -829,8 +831,7 @@ def jvp(f: DifferentiableLike, params, tangents, *, method=None, h=None, argnum=
     >>> x = jnp.array([0.1, 0.2])
     >>> tangent = jnp.array([0.3, 0.6])
     >>> jvp(x, tangent)
-    [array([0.09983342, 0.04      , 0.02      ]),
-    array([0.29850125, 0.24000006, 0.12      ])]
+    (array([0.09983342, 0.04      , 0.02      ]), array([0.29850125, 0.24      , 0.12      ]))
 
     **Example 2 (argnum usage)**
 
@@ -854,7 +855,7 @@ def jvp(f: DifferentiableLike, params, tangents, *, method=None, h=None, argnum=
     >>> params = jnp.array([[0.54, 0.3154], [0.654, 0.123]])
     >>> dy = jnp.array([[1.0, 1.0], [1.0, 1.0]])
     >>> workflow(params, dy)
-    [array(0.78766064), array(-0.7011436)]
+    (array(0.78766064), array(-0.7011436))
     """
 
     def check_is_iterable(x, hint):
@@ -873,15 +874,14 @@ def jvp(f: DifferentiableLike, params, tangents, *, method=None, h=None, argnum=
 
         results = jvp_p.bind(*params, *tangents, jaxpr=jaxpr, fn=fn, grad_params=grad_params)
 
-        if out_tree.children() != []:
-            midpoint = len(results) // 2
-            func_res = results[:midpoint]
-            jvps = results[midpoint:]
-            func_res = tree_unflatten(out_tree, func_res)
-            jvps = tree_unflatten(out_tree, jvps)
-            results = tuple([func_res, jvps])
-        else:
-            results = tuple(results)
+        midpoint = len(results) // 2
+        func_res = results[:midpoint]
+        jvps = results[midpoint:]
+
+        func_res = tree_unflatten(out_tree, func_res)
+        jvps = tree_unflatten(out_tree, jvps)
+        results = (func_res, jvps)
+
     else:
         results = jax.jvp(f, params, tangents)
 
@@ -894,7 +894,8 @@ def vjp(f: DifferentiableLike, params, cotangents, *, method=None, h=None, argnu
 
     This function allows the Vector-Jacobian Product of a hybrid quantum-classical function to be
     computed within the compiled program. Outside of a compiled function, this function will simply
-    dispatch to its JAX counterpart ``jax.vjp``.
+    dispatch to its JAX counterpart ``jax.vjp``. The function ``f`` can return any pytree-like
+    shape.
 
     Args:
         f(Callable): Function-like object to calculate JVP for
@@ -907,8 +908,8 @@ def vjp(f: DifferentiableLike, params, cotangents, *, method=None, h=None, argnu
         h (float): the step-size value for the finite-difference (``"fd"``) method
         argnum (Union[int, List[int]]): the params' indices to differentiate.
 
-    Returns (Tuple[Array]):
-        Return values of ``f`` paired with the JVP values.
+    Returns:
+        Tuple[Any]): Return values of ``f`` paired with the VJP values.
 
     Raises:
         TypeError: invalid parameter types
@@ -929,8 +930,7 @@ def vjp(f: DifferentiableLike, params, cotangents, *, method=None, h=None, argnu
     >>> x = jnp.array([0.1, 0.2])
     >>> dy = jnp.array([-0.5, 0.1, 0.3])
     >>> vjp(x, dy)
-    [array([0.09983342, 0.04      , 0.02      ]),
-    array([-0.43750208,  0.07000001])]
+    (array([0.09983342, 0.04      , 0.02      ]), (array([-0.43750208,  0.07      ]),))
     """
 
     def check_is_iterable(x, hint):
@@ -947,19 +947,19 @@ def vjp(f: DifferentiableLike, params, cotangents, *, method=None, h=None, argnu
 
         jaxpr, out_tree = _make_jaxpr_check_differentiable(fn, grad_params, *params)
 
+        cotangents, _ = tree_flatten(cotangents)
+
         results = vjp_p.bind(*params, *cotangents, jaxpr=jaxpr, fn=fn, grad_params=grad_params)
 
-        if out_tree.children() != []:
-            func_res = results[: len(jaxpr.out_avals)]
-            vjps = results[len(jaxpr.out_avals) :]
-            func_res = tree_unflatten(out_tree, func_res)
-            # We do not change the shape of the VJP as it is the same as the parameters not
-            # as the output
-            results = tuple([func_res, tuple(vjps)])
-        else:
-            func_res = results[: len(jaxpr.out_avals)]
-            vjps = results[len(jaxpr.out_avals) :]
-            results = tuple([*func_res, tuple(vjps)])
+        func_res = results[: len(jaxpr.out_avals)]
+        vjps = results[len(jaxpr.out_avals) :]
+        func_res = tree_unflatten(out_tree, func_res)
+        # We do not change the shape of the VJP as it is the same as the parameters not
+        # as the output
+        # TODO: update when we add support for pytrees params
+
+        results = tuple([func_res, tuple(vjps)])
+
     else:
         primal_outputs, vjp_fn = jax.vjp(f, *params)
 
@@ -1147,9 +1147,7 @@ class MidCircuitMeasure(HybridOp):
         op = self
         wire = op.in_classical_tracers[0]
         qubit = qrp.extract([wire])[0]
-
-        # Check if the postselect value was given, otherwise default to None
-        postselect = op.in_classical_tracers[1] if len(op.in_classical_tracers) > 1 else None
+        postselect = op.in_classical_tracers[1]
 
         qubit2 = op.bind_overwrite_classical_tracers(ctx, trace, qubit, postselect=postselect)
         qrp.insert([wire], [qubit2])
@@ -1954,7 +1952,9 @@ def while_loop(cond_fn):
     return _body_query
 
 
-def measure(wires, postselect: Optional[int] = None) -> DynamicJaxprTracer:
+def measure(
+    wires, reset: Optional[bool] = False, postselect: Optional[int] = None
+) -> DynamicJaxprTracer:
     """A :func:`qjit` compatible mid-circuit measurement for PennyLane/Catalyst.
 
     .. important::
@@ -1964,6 +1964,7 @@ def measure(wires, postselect: Optional[int] = None) -> DynamicJaxprTracer:
 
     Args:
         wires (Wires): The wire of the qubit the measurement process applies to
+        reset (Optional[bool]): Whether to reset the wire to the |0⟩ state after measurement.
         postselect (Optional[int]): Which basis state to postselect after a mid-circuit measurement.
 
     Returns:
@@ -2010,6 +2011,22 @@ def measure(wires, postselect: Optional[int] = None) -> DynamicJaxprTracer:
 
     >>> circuit()
     -1.0
+
+    **Example with reset**
+
+    .. code-block:: python
+
+        dev = qml.device("lightning.qubit", wires=1)
+
+        @qjit
+        @qml.qnode(dev)
+        def circuit():
+            qml.Hadamard(0)
+            m = measure(0, reset=True)
+            return qml.expval(qml.PauliZ(0))
+
+    >>> circuit()
+    1.0
     """
     EvaluationContext.check_is_tracing("catalyst.measure can only be used from within @qjit.")
     EvaluationContext.check_is_quantum_tracing(
@@ -2020,22 +2037,31 @@ def measure(wires, postselect: Optional[int] = None) -> DynamicJaxprTracer:
     if len(wires) != 1:
         raise TypeError(f"One classical argument (a wire) is expected, got {wires}")
 
-    in_classical_tracers = wires
+    # Copy, so wires remain unmodified
+    in_classical_tracers = wires.copy()
 
-    # Check the postselect value. If given, add it to the classical tracers list
-    if postselect is not None:
-        if postselect not in [0, 1]:
-            raise TypeError(f"postselect must be '0' or '1', got {postselect}")
-        in_classical_tracers.append(postselect)
+    if postselect is not None and postselect not in [0, 1]:
+        raise TypeError(f"postselect must be '0' or '1', got {postselect}")
+    in_classical_tracers.append(postselect)
 
     # assert len(ctx.trace.frame.eqns) == 0, ctx.trace.frame.eqns
-    out_classical_tracer = new_inner_tracer(ctx.trace, get_aval(True))
+    m = new_inner_tracer(ctx.trace, get_aval(True))
     MidCircuitMeasure(
         in_classical_tracers=in_classical_tracers,
-        out_classical_tracers=[out_classical_tracer],
+        out_classical_tracers=[m],
         regions=[],
     )
-    return out_classical_tracer
+
+    # If reset was requested, reset qubit only if the measurement result was 1
+    if reset:
+
+        @cond(m)
+        def reset_fn():
+            qml.PauliX(wires=wires)
+
+        reset_fn()
+
+    return m
 
 
 def adjoint(f: Union[Callable, Operator]) -> Union[Callable, Operator]: